Login Email And Password
Conditional Statements:

The first if statement checks if Login is "fahadkhann@gmail.com" and Password is "12345". If true, it prints a success message.
The second else if statement checks if Login is "fahadkhann@gmail.com" and Password is not "12345". If true, it prints a message about the correct Login but incorrect Password.
The third else if statement checks if Login is not "fahadkhann@gmail.com" and Password is "123456". If true, it prints a message about the incorrect Login but correct Password.
The else statement is a catch-all for cases where none of the above conditions are met. It prints "Try Again."